+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ### Added
 
+- Add a `dune-release config` subcommand to display and edit the global
+  configuration (#220, @NathanReb).
 - Add command `delegate-info` to print information needed by external
   release scripts (#221, @pitag-ha)
 - Use Curly instead of Cmd to interact with github (#202, @gpetiot)